refactor(package): 폴더링 변경
