+++ app/src/main/java/org/android/go/sopt/util/base/BindingFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.util
+package org.android.go.sopt.util.base
 
 import android.os.Bundle
 import android.view.LayoutInflater
